misc: add some comments
